refactor: sections into directories
