+++ apps/testnet-leaderboard/components/FAQSection/index.tsx
