--- apps/testnet-leaderboard/components/tables/index.ts
@@ -1,3 +1,0 @@
-export { default as AddressCell } from './AddressCell';
-export { default as BadgesCell } from './BadgesCell';
-export { default as HeaderCell } from './HeaderCell';
